conflicts stash

+++ ex1/chess/chessMapUtils.c
@@ -115,7 +115,7 @@ MapKeyElement stringCopyFunc (MapKeyElement to_copy){
     if(!to_copy){
         return NULL;
     }
-    int length = strlen((char*)to_copy);
+    int length = (int)strlen((char*)to_copy);
     char* to_return = malloc(length+1);
     if(!to_return){
         return NULL;
